requirements.txt
fastapi[all]==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
pandas==1.5.3
psycopg2==2.9.5
fastapi-utils==0.2.1